requirements.txt
ttkbootstrap==1.13.10
openai-whisper